streamlit/requirements.txt
pandas==1.5.3
pymongo==4.3.3
python-dotenv==0.21.1
streamlit==1.18.1
